+++ dev_kit_mcp_server/core/__init__.py
@@ -1,6 +1,6 @@
 """core repo operations."""
 
-from dev_kit_mcp_server.tools.core.base_op import AsyncOperation
+from .base_op import AsyncOperation
 
 __all__ = [
     "AsyncOperation",
