+++ dev_kit_mcp_server/tools/file_sys/edit.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict
 
-from ..core import AsyncOperation
+from ...core import AsyncOperation
 
 
 @dataclass
