+++ dev_kit_mcp_server/tools/git/status.py
@@ -3,7 +3,7 @@
 from dataclasses import dataclass
 from typing import Any, Dict
 
-from ..core import AsyncOperation
+from ...core import AsyncOperation
 
 
 @dataclass
